Move `CurrentDateTimeSource` to `core/data/time` package
