+++ core/data/src/commonMain/kotlin/dev/sasikanth/rss/reader/data/time/CurrentDateTimeSource.kt
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.sasikanth.rss.reader.utils
+package dev.sasikanth.rss.reader.data.time
 
 import dev.sasikanth.rss.reader.di.scopes.AppScope
-import kotlin.time.Duration.Companion.hours
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
@@ -30,6 +29,7 @@ import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
 import me.tatarka.inject.annotations.Inject
+import kotlin.time.Duration.Companion.hours
 
 @Inject
 @AppScope
