+++ shared/src/androidMain/kotlin/dev/sasikanth/rss/reader/di/ApplicationComponent.kt
@@ -22,8 +22,8 @@ import dev.sasikanth.rss.reader.app.AppInfo
 import dev.sasikanth.rss.reader.data.repository.RssRepository
 import dev.sasikanth.rss.reader.data.repository.SettingsRepository
 import dev.sasikanth.rss.reader.data.sync.SyncCoordinator
+import dev.sasikanth.rss.reader.data.time.CurrentDateTimeSource
 import dev.sasikanth.rss.reader.di.scopes.AppScope
-import dev.sasikanth.rss.reader.utils.CurrentDateTimeSource
 import me.tatarka.inject.annotations.Component
 import me.tatarka.inject.annotations.Provides
